+++ src/main/java/com/lshzzz/mato/model/room/dto/ParticipantReadyRequest.java
@@ -1,3 +1,3 @@
-package com.lshzzz.mato.controller.rooms;
+package com.lshzzz.mato.model.room.dto;
 
 public record ParticipantReadyRequest(String nickname, boolean ready) {}
